+++ src/OpenRiaServices.Hosting.AspNetCore/Framework/AspNetCore/Operations/OperationInvoker.cs
@@ -72,22 +72,30 @@ namespace OpenRiaServices.Hosting.AspNetCore.Operations
         protected async Task<(ServiceQuery, object[])> ReadParametersFromBodyAsync(HttpContext context)
         {
             var request = context.Request;
-            var contentLength = request.ContentLength;
-            if (!contentLength.HasValue || contentLength < 0 || contentLength > int.MaxValue)
-                throw new BadHttpRequestException("invalid lenght", (int)System.Net.HttpStatusCode.LengthRequired);
 
-            var length = (int)contentLength;
-            var buffer = ArrayPool<byte>.Shared.Rent(length);
+            int initialCapacity = request.ContentLength switch
+            {
+                long contentLength and >= 0 and <= int.MaxValue => Math.Min((int)contentLength, 4096),
+                null => 4096,
+                _ => throw new BadHttpRequestException("invalid lenght", (int)System.Net.HttpStatusCode.BadRequest)
+            };
+
+            // To prevent DOS attacks where an attacker can allocate arbitary large memory by setting content-length to a large value
+            // We only allocate a maximum of 4K directly
+            using var ms = new ArrayPoolStream(ArrayPool<byte>.Shared, maxBlockSize: 4 * 1024 * 1024);
+            ms.Reset(initialCapacity); // Initial capacity up to 4K
+
+            await request.BodyReader.CopyToAsync(ms).ConfigureAwait(false);
+            ArraySegment<byte> memory = ms.GetRentedArrayAndClear();
+
             try
             {
-                await CopyToBufferAsync(request, buffer, length).ConfigureAwait(false);
-
-                using var reader = XmlDictionaryReader.CreateBinaryReader(buffer, 0, length, null, XmlDictionaryReaderQuotas.Max);
-                return ReadParametersFromBody(reader);
+                using var reader = BinaryMessageReader.Rent(memory);
+                return ReadParametersFromBody(reader.XmlDictionaryReader);
             }
             finally
             {
-                ArrayPool<byte>.Shared.Return(buffer);
+                ArrayPool<byte>.Shared.Return(memory.Array);
             }
         }
 
@@ -255,7 +263,7 @@ namespace OpenRiaServices.Hosting.AspNetCore.Operations
                 }
             }
 
-            return WriteError(context, new DomainServiceFault { OperationErrors = errors, ErrorCode = 422 });
+            return WriteError(context, new DomainServiceFault { OperationErrors = errors, ErrorCode = StatusCodes.Status422UnprocessableEntity });
         }
 
 
@@ -379,21 +387,5 @@ namespace OpenRiaServices.Hosting.AspNetCore.Operations
             domainService.Initialize(serviceContext);
             return domainService;
         }
-
-
-        private static async Task CopyToBufferAsync(HttpRequest request, byte[] buffer, int length)
-        {
-            int offset = 0;
-            var body = request.Body;
-            do
-            {
-                int read = await body.ReadAsync(buffer, offset, length - offset).ConfigureAwait(false);
-                offset += read;
-
-                if (read == 0)
-                    throw new BadHttpRequestException(InvalidContentMessage);
-            } while (offset < length);
-        }
-
     }
 }